on_timer: Do not pass an exception up the calling chain
In the current implementation of the timer helper, it immediately detaches
itself from the event loop if its callback raise an unhandled exception
(so no processing occurs thereafter).

Since ConcatFilter.on_timer() does not catch any exception at all,
there's a quite good chance that the flushing thread stops working
in the middle of operation with the following error:

> [error]: #0 Timer detached. title=:filter_concat_timer

This patch is the quick fix for this issue.