Handle partial CRI-O log format on CRI-O specific mode
In the newer kubernetes, it start to use CRI-O format for logging.
We should handle this format on concat plugin with specific mode.

ref: https://github.com/kubernetes/community/blob/master/contributors/design-proposals/node/kubelet-cri-logging.md

On kubelet implementation for parsing, stream type is just for checking
stream names validity and then they are thrown away.
Users want to obtain concatenated log contents not for attached metadata.

ref: https://github.com/kubernetes/kubernetes/blob/629d5ab21349021cf7d38236620785071ee541b4/pkg/kubelet/kuberuntime/logs/logs.go#L124-L168

When we use regexp to concat CRI-O log, it is hard to remove attached
metadata records such as stream names(stdout|stderr) and partial_flag(P
or F).

Signed-off-by: Hiroshi Hatake <hatake@calyptia.com>